// client/src/pages/order-management/types/order.types.ts
export type TransactionMode = 'Physical' | 'Email' | 'Telephone';
// ...
export interface TransactionModeData {
  mode: TransactionMode;
  email?: string;
  phoneNumber?: string;
  physicalAddress?: string;
  euin?: string;
}